Kafka egress connector tests.
